✨ don't print veneer symbols
Veneers might be inserted for any of the floating point function
symbols, depending on memory layout and placement. It's extra noise in
the output that's not particularly useful, so set `grep` to only match
on whole words for the symbol names.

Added test coverage by putting a hardcoded symbol with the veneer naming
style, and verify there's no match.

Fixes #2.